# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Learning Management System API
  description: LMS API for managing courses, modules, lessons, enrollments and progress tracking
  version: 1.0.0
  contact:
    name: LMS Team
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8081/api/v1
    description: Development
  - url: /api/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ModuleCreateRequest:
      type: object
      required:
        - courseId
        - moduleTitle
        - sequence
      properties:
        courseId:
          type: integer
          format: int64
          example: 1
        moduleTitle:
          type: string
          maxLength: 50
          minLength: 1
          example: Spring Core
        moduleDescription:
          type: string
          maxLength: 100
        sequence:
          type: integer
          minimum: 1
          example: 1
    ModuleUpdateRequest:
      type: object
      properties:
        moduleTitle:
          type: string
          maxLength: 50
          minLength: 1
        moduleDescription:
          type: string
          maxLength: 100
        sequence:
          type: integer
          minimum: 1
        isActive:
          type: boolean
    ModuleResponse:
      type: object
      properties:
        moduleId:
          type: integer
          format: int64
        courseId:
          type: integer
          format: int64
        moduleTitle:
          type: string
        moduleDescription:
          type: string
        sequence:
          type: integer
        lessons:
          type: array
          items:
            $ref: './openapi/common.yaml#/components/schemas/LessonSummary'
          default: []
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ModuleListResponse:
      type: object
      properties:
        message:
          type: string
        data:
          type: array
          items:
            $ref: './openapi/common.yaml#/components/schemas/ModuleSummary'
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
        timestamp:
          type: string
          format: date-time
...
    ProgressUpdateRequest:
      type: object
      required:
        - lessonStatus
      properties:
        lessonStatus:
          $ref: './openapi/common.yaml#/components/schemas/LessonStatus'
          example: STARTED
    ProgressResponse:
      type: object
      properties:
        progressId:
          type: integer
          format: int64
        lessonId:
          type: integer
          format: int64
        lessonTitle:
          type: string
        userId:
          type: integer
          format: int64
        moduleId:
          type: integer
          format: int64
        courseId:
          type: integer
          format: int64
        lessonStatus:
          $ref: './openapi/common.yaml#/components/schemas/LessonStatus'
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
    ProgressListResponse:
      type: object
      properties:
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProgressResponse'
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
        timestamp:
          type: string
          format: date-time
security:
  - bearerAuth: []
paths:
...
  /modules:
    post:
      tags:
        - Modules
      summary: Create a new module
      operationId: createModule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleCreateRequest'
      responses:
        '201':
          description: Module created successfully
          content:
            application/json:
              schema:
                type: object
                title: ModuleCreateResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/ModuleResponse'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Not found
        '409':
          description: Conflict

    get:
      tags:
        - Modules
      summary: List modules
      operationId: listModules
      parameters:
        - name: courseId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: active
          in: query
          schema:
            type: boolean
            default: true
        - name: pageNo
          in: query
          required: true
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            default: sequence
        - name: sortOrder
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
      responses:
        '200':
          description: Modules fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleListResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
  /modules/{moduleId}:
    get:
      tags:
        - Modules
      summary: Get module by ID
      operationId: getModule
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: includeLessons
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Module fetched successfully
          content:
            application/json:
              schema:
                type: object
                title: ModuleGetResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/ModuleResponse'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Not found

    put:
      tags:
        - Modules
      summary: Update module
      operationId: updateModule
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleUpdateRequest'
      responses:
        '200':
          description: Module updated successfully
          content:
            application/json:
              schema:
                type: object
                title: ModuleUpdateResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/ModuleResponse'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Not found
        '409':
          description: Conflict

    delete:
      tags:
        - Modules
      summary: Delete module
      operationId: deleteModule
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Module deleted successfully
          content:
            application/json:
              schema:
                type: object
                title: ModuleDeleteResponse
                properties:
                  message:
                    type: string
                  data:
                    type: object
                  timestamp:
                    type: string
                    format: date-time
        '403':
          description: Forbidden
        '404':
          description: Not found
        '409':
          description: Conflict
...
  /progress:
    get:
      tags:
        - Progress
      summary: List progress records
      operationId: listProgress
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: courseId
          in: query
          schema:
            type: integer
            format: int64
        - name: moduleId
          in: query
          schema:
            type: integer
            format: int64
        - name: lessonStatus
          in: query
          schema:
            type: string
            enum:
              - UNSTARTED
              - STARTED
              - FINISHED
        - name: pageNo
          in: query
          required: true
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            default: lessonId
        - name: sortOrder
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
      responses:
        '200':
          description: Progress records fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressListResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
  /progress/{progressId}:
    get:
      tags:
        - Progress
      summary: Get progress by ID
      operationId: getProgress
      parameters:
        - name: progressId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Progress fetched successfully
          content:
            application/json:
              schema:
                type: object
                title: ProgressGetResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/ProgressResponse'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Not found

    put:
      tags:
        - Progress
      summary: Update progress
      operationId: updateProgress
      parameters:
        - name: progressId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgressUpdateRequest'
      responses:
        '200':
          description: Progress updated successfully
          content:
            application/json:
              schema:
                type: object
                title: ProgressUpdateResponse
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/ProgressResponse'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Not found
        '409':
          description: Conflict